refactor(logSink): share the tee and bounded writer, and explain a disabled stream
The daemon sink carries the same tee pattern with the same asynchronous
blind spot, so the transport primitives move to src/shared/logTransport.ts
where the daemon can reach them without importing electron. The main sink
re-exports them, so existing call sites are unchanged.

Two fixes ride along:

- A disabled pass-through now writes one line to the file naming the stream
  and the error code. Any stream error still disables forwarding — narrowing
  the set risks a code we failed to anticipate restarting the storm — but
  silence with no trace made a dead console indistinguishable from a process
  that simply stopped logging.
- isBrokenPipeError also matches EBADF, which reaches the same loop when the
  fd is closed underneath the process.
